+++ waspc/e2e-test/test-outputs/waspComplexTest-golden/waspComplexTest/.wasp/out/sdk/wasp/server/auth/oauth/index.ts
@@ -1,6 +1,9 @@
 // PUBLIC API
 export { google } from './providers/google.js';
 
+
+
+
 // PRIVATE API
 export {
   loginPath,